Fix bug in SendKey function
The sendkey function assumes that Go's uint type is the same length as
C's uint type. This is not true on many systems, especially x86_64
Linux, where Go's uint is usually 8 bytes, but C's is usually 4.

This change addresses the issue by explicitly converting the slice of Go
keycodes to a slice of C keycodes.